requirements.txt
python-libinput
mycroft-messagebus-client
